+++ changelog.md
@@ -5,6 +5,7 @@
 ### Patch
 
 - Updated dev dependencies.
+- Use [`test-director`](https://npm.im/test-director) as a dev dependency for tests.
 - Improved the installation instructions in the readme.
 - Remove [`node-fetch`](https://npm.im/node-fetch) from code examples. Modern Node.js supports the `fetch` global.